cmake: Fix build on 10.4.
The 10.4 build was failing due to an attempt to use lutimes(), which
didn't exist until 10.5.  There's another portion of the cmake code
that includes a configure test for lutimes(), but it doesn't apply to
this area.  The fix is to add an OS version check to the existing
conditionals.

Checks of this form should more correctly use MIN_REQUIRED rather than
MAX_ALLOWED, but this uses the latter for consistency with the other
places, both in the patches and upstream.

Since this only fixes broken builds, there is no need for a revbump.

TESTED:
Built on 10.4-10.5 ppc, 10.4-10.6 i386, and 10.6-10.15 x86_64.
Did not test variants, due to dependency complications, and the low
likelihood of an interaction with this change.
Built libaes_siv as a test case.